Add TargetableActionNode for actions that can have multiple targets
Some actions might have multiple targets that they can be performed on. For
example, a soldier may be able to attack 3 different enemy soldiers.

The TargetableActionNode can be updated with information about available
targets. These stats will be used to calculate the utility of performing the
action on each target. A target will be selected using the specified selection
method, and then the utility of the action will be the utility of performing the
action on the chosen target.